Fix memory leak / incorrect refcount decrement in lua bindings.
References to values are looked up before removal, for GC purposes.
This had the potential to return the value associated with a longer prefix.
If both this prefix and the target prefix were present, the wrong refcount
would be decremented.

Additionally, references were overwritten when calling :insert with a prefix
that already had a value associated with it.

This changes the code to use lpm_retrieve which does an exact match,
and then decrefs as appropriate.